Refactor bots to agents folder
