+++ agents/bots/README.md
