+++ agents/bots/gang/index.ts
