+++ agents/bots/key-check/handlers/debug.ts
