+++ agents/bots/key-check/index.ts
