+++ agents/bots/utils/atttachment.ts
