+++ agents/bots/utils/inline-actions/types/ActionsContent.ts
