+++ agents/bots/utils/inline-actions/types/IntentContent.ts
