+++ agents/bots/utils/usdc.ts
