+++ agents/version-management/cli-node-versions.ts
@@ -1,7 +1,7 @@
 import { execSync } from "child_process";
 import * as fs from "node:fs";
 import * as path from "node:path";
-import { VersionList } from "version-management/sdk-node-versions";
+import { VersionList } from "./sdk-node-versions";
 
 function showHelp() {
   console.log(`
